+++ cardano-testnet/src/Testnet/Util/Runtime.hs
@@ -2,6 +2,8 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DuplicateRecordFields #-}
 {-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module Testnet.Util.Runtime
   ( LeadershipSlot(..)
@@ -19,16 +21,24 @@ module Testnet.Util.Runtime
   , poolNodeStdout
   , readNodeLoggingFormat
   , startNode
+  , ShelleyGenesis(..)
+  , shelleyGenesis
+  , getStartTime
+  , fromNominalDiffTimeMicro
   ) where
 
 import           Prelude
 
 import           Control.Monad
-import           Data.Aeson (FromJSON)
+import           Control.Monad.Error.Class
+import           Control.Monad.IO.Class
+import           Control.Monad.Trans.Except
+import qualified Data.Aeson as A
 import qualified Data.List as L
-
 import           Data.Text (Text)
+import           Data.Time.Clock (UTCTime)
 import           GHC.Generics (Generic)
+import           GHC.Stack
 import qualified Hedgehog as H
 import           Hedgehog.Extras.Stock.IO.Network.Sprocket (Sprocket (..))
 import qualified Hedgehog.Extras.Stock.IO.Network.Sprocket as IO
@@ -36,12 +46,19 @@ import qualified Hedgehog.Extras.Stock.String as S
 import qualified Hedgehog.Extras.Test.Base as H
 import qualified Hedgehog.Extras.Test.File as H
 import qualified Hedgehog.Extras.Test.Process as H
-
 import           System.FilePath.Posix ((</>))
 import qualified System.Info as OS
 import qualified System.IO as IO
 import qualified System.Process as IO
 
+import           Cardano.Api
+import qualified Cardano.Chain.Genesis as G
+import           Cardano.Crypto.ProtocolMagic (RequiresNetworkMagic (..))
+import           Cardano.Ledger.Crypto (StandardCrypto)
+import           Cardano.Ledger.Shelley.Genesis
+import           Cardano.Node.Configuration.POM
+import qualified Cardano.Node.Protocol.Byron as Byron
+import           Cardano.Node.Types
 import qualified Testnet.Util.Process as H
 
 data NodeLoggingFormat = NodeLoggingFormatAsJson | NodeLoggingFormatAsText deriving (Eq, Show)
@@ -107,6 +124,37 @@ bftSprockets = fmap nodeSprocket . bftNodes
 
 poolSprockets :: TestnetRuntime -> [Sprocket]
 poolSprockets = fmap (nodeSprocket . poolRuntime) . poolNodes
+
+shelleyGenesis :: (H.MonadTest m, MonadIO m, HasCallStack) => TestnetRuntime -> m (ShelleyGenesis StandardCrypto)
+shelleyGenesis TestnetRuntime{shelleyGenesisFile} = withFrozenCallStack $
+  H.evalEither =<< H.evalIO (A.eitherDecodeFileStrict' shelleyGenesisFile)
+
+getStartTime :: (H.MonadTest m, MonadIO m, HasCallStack) => FilePath -> TestnetRuntime -> m UTCTime
+getStartTime tempRootPath TestnetRuntime{configurationFile} = withFrozenCallStack $ H.evalEither <=< H.evalIO . runExceptT $ do
+  byronGenesisFile <-
+    decodeNodeConfiguration configurationFile >>=
+      \case
+        NodeProtocolConfigurationCardano NodeByronProtocolConfiguration{npcByronGenesisFile} _ _ _ _ ->
+          pure $ unGenesisFile npcByronGenesisFile
+        NodeProtocolConfigurationByron NodeByronProtocolConfiguration{npcByronGenesisFile} ->
+          pure $ unGenesisFile npcByronGenesisFile
+        unsupported ->
+          throwError $ unwords
+            [ "cannot find byron configuration path in"
+            , configurationFile
+            , "- found instead:"
+            , show unsupported
+            ]
+  let byronGenesisFilePath = tempRootPath </> byronGenesisFile
+  G.gdStartTime . G.configGenesisData <$> decodeGenesisFile byronGenesisFilePath
+  where
+    decodeNodeConfiguration :: FilePath -> ExceptT String IO NodeProtocolConfiguration
+    decodeNodeConfiguration file = do
+      partialNodeCfg <- ExceptT $ A.eitherDecodeFileStrict' file
+      fmap ncProtocolConfig . liftEither . makeNodeConfiguration $ defaultPartialNodeConfiguration <> partialNodeCfg
+    decodeGenesisFile :: FilePath -> ExceptT String IO G.Config
+    decodeGenesisFile fp = withExceptT displayError $
+      Byron.readGenesis (GenesisFile fp) Nothing RequiresNoMagic
 
 readNodeLoggingFormat :: String -> Either String NodeLoggingFormat
 readNodeLoggingFormat = \case